refactor: Separate content types
