+++ changes/216.changed
@@ -1,0 +1,1 @@
+Separated content types definitions to distinct module.
